Replace antlr2 grammer with antlr4
The lexer is essentially unchanged, just tweaked to satisfy antlr4 expectations. The order of rules was modified for
readability, but is functionally the same.

The parser was subtantially changed, both to satisfy antlr4 as well as to take advantage of the new LL(*) capabilities.

There is a new AST as antlr4 does not provide its own AST. The tree structure is identical to the old AST, but is
immutable and strongly typed. The evaluation of the new AST is identical to that of the old AST.

Aside from those using unicode, all test cases pass on both the antlr2 and antlr4 implementations. The unicode test
cases now preserve the unicode characters in the results instead of mapping them to U+003F but are otherwise unchanged.