Replace antlr2 grammer with antlr4
The lexer is essentially unchanged, just tweaked to satisfy antlr4 expectations. The order of rules was modified for
readability, but is functionally the same.

The parser was subtantially changed, both to satisfy antlr4 as well as to take advantage of the new LL(*) capabilities.

There is a new strongly-typed AST as antlr4 does not provide its own AST. The parser test cases had to have minor tweaks
made to account for the differences:
- there is no more distinction between single- and double-quoted strings internally
- hex-coded integers are now the same as all other integers internally
- unicode strings are now supported

Meanwhile, all evaluation tests cases remain valid, and new multibyte unicode cases have been added.

Reformat lexer

Remove unused tokens, QUOTE and DID