Match end of input so we do not leave trailing inputs
This forces parse errors in cases that are obviously not correct (e.g., assigning to a string literal). This requires
the lexer to consume all tokens and not allow dropping them. It also require explicitly matching multiple expressions in
a single line, whereas before the parse would simply stop after the first expression and return it.

In order to report unconsume input, we install a `BailErrorStrategy` which - unlike the default - does not attempt error
recovery.